services/visual_generator/requirements.txt
flask==2.0.1
python-dotenv==0.19.0
firebase-admin==5.3.0
requests==2.26.0
gunicorn==20.1.0
python-jose==3.3.0
Pillow==8.3.2 